Simplify for shinyslack usage. (#60)
* Simplify for shinyslack usage.

Had to deal with the change of not directly activating the team, may wish to look into ways to deal with that more easily.

* Bring back login check.

Oops! The server is the same for all iterations! Consider finding a way to change that in the future so things don't try to run until they should work.